requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via httpx
attrs==25.4.0
    # via aiohttp
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
filetype==1.2.0
    # via langchain-google-genai
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
google-ai-generativelanguage==0.9.0
    # via
    #   -r requirements.in
    #   langchain-google-genai
google-api-core[grpc]==2.28.1
    # via google-ai-generativelanguage
google-auth==2.42.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==2.1.0
    # via pytest
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.0.3
    # via -r requirements.in
langchain-core==1.0.3
    # via
    #   langchain
    #   langchain-google-genai
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-google-genai==3.0.1
    # via -r requirements.in
langgraph==1.0.2
    # via langchain
langgraph-checkpoint==3.0.0
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.2
    # via langgraph
langgraph-sdk==0.2.9
    # via langgraph
langsmith==0.4.40
    # via langchain-core
loguru==0.7.3
    # via -r requirements.in
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
orjson==3.11.4
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.11.0
    # via langgraph-checkpoint
packaging==24.2
    # via
    #   langchain-core
    #   langsmith
    #   pytest
pluggy==1.5.0
    # via pytest
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.8
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.3
    # via
    #   langchain
    #   langchain-core
    #   langchain-google-genai
    #   langgraph
    #   langsmith
pydantic-core==2.41.4
    # via pydantic
pytest==8.3.5
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   langchain-core
requests==2.32.5
    # via
    #   google-api-core
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
rsa==4.9.1
    # via google-auth
sniffio==1.3.1
    # via anyio
tenacity==8.5.0
    # via langchain-core
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   grpcio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
urllib3==2.5.0
    # via requests
xxhash==3.6.0
    # via langgraph
yarl==1.22.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
